split tools api commands into seperate files
